# Add tag interning update interning and use new methods