--- a/Gopkg.lock
+++ b/Gopkg.lock
@@ -732,11 +732,11 @@
 
 [[projects]]
   branch = "master"
-  digest = "1:73fbe6a6838e8d312f8d78fce7185282f409d35e27906027bd4f5ccf7a5d3ae9"
+  digest = "1:377ce35e894527980829ad71a7eb04380d94394bce8240d69269825e350ca222"
   name = "github.com/robert-milan/go-object-interning"
   packages = ["."]
   pruneopts = "NUT"
-  revision = "f911bee7f9bef4c3d0bb590205eb58f3259820df"
+  revision = "d94daf92a88dce589b0aa1e920ae199677101ff4"
 
 [[projects]]
   digest = "1:c5741d3d03a06220bcca801547f28de803103249b739eb5537e54b77e89f435a"