… vendor/github.com/robert-milan/go-object-interning/object_intern.go