This seems to lead to a significant improvement in ingest performance. Note that in this benchmark `1` operation is ingesting `100` metric points. So if the benchmark says that an operation takes `228ms` this means ingesting `100` points takes `228ms`.

Master:
```
BenchmarkProcessMetricPoint-8             300000             53073 ns/op             644 B/op          4 allocs/op
```
With Interning, without this PR:
```
BenchmarkProcessMetricPoint-8             100000            228065 ns/op             675 B/op          7 allocs/op
```
With Interning, with this PR:
```
BenchmarkProcessMetricPoint-8             200000             70658 ns/op             660 B/op          5 allocs/op
```