@robert-milan 2019-07-26
> What does "Add tag interning update interning" mean ?

I think it means "update the interning library"