@replay 2019-07-26
sorry, that title was generated from the commit and dropped a comment. was supposed to be "Add tag interning, update interning and use new methods"